moved engine to it's own file and added easy export function
